Fix doctests.

+++ sparse/numba_backend/_sparse_array.py
@@ -689,7 +689,7 @@ class SparseArray:
         mean along all axes.
 
         >>> s.mean()
-        0.5
+        np.float64(0.5)
         """
 
         if axis is None:
@@ -769,7 +769,7 @@ class SparseArray:
         variance along all axes.
 
         >>> s.var()
-        0.5
+        np.float64(0.5)
         """
         axis = normalize_axis(axis, self.ndim)
